fix python <3.8
