+++ python/.gitignore
@@ -4,4 +4,5 @@ taichi/include
 taichi/examples
 taichi/assets
 taichi/tests
+taichi/tests38
 release
