+++ setup.py
@@ -74,6 +74,7 @@ def remove_tmp(taichi_dir):
     shutil.rmtree(os.path.join(taichi_dir, 'assets'), ignore_errors=True)
     shutil.rmtree(os.path.join(taichi_dir, 'examples'), ignore_errors=True)
     shutil.rmtree(os.path.join(taichi_dir, 'tests'), ignore_errors=True)
+    shutil.rmtree(os.path.join(taichi_dir, 'tests38'), ignore_errors=True)
 
 
 class CMakeExtension(Extension):
@@ -87,6 +88,7 @@ class EggInfo(egg_info):
         remove_tmp(taichi_dir)
 
         shutil.copytree('tests/python', os.path.join(taichi_dir, 'tests'))
+        shutil.copytree('tests/python38', os.path.join(taichi_dir, 'tests38'))
         shutil.copytree('examples', os.path.join(taichi_dir, 'examples'))
         shutil.copytree('external/assets', os.path.join(taichi_dir, 'assets'))
 
@@ -208,7 +210,7 @@ class Clean(clean):
             remove_tree(self.build_temp, dry_run=self.dry_run)
         generated_folders = ('bin', 'dist', 'python/taichi/assets',
                              'python/taichi/lib', 'python/taichi/examples',
-                             'python/taichi/tests', 'python/taichi.egg-info')
+                             'python/taichi/tests', 'python/taichi/tests38', 'python/taichi.egg-info')
         for d in generated_folders:
             if os.path.exists(d):
                 remove_tree(d, dry_run=self.dry_run)
