+++ tests/python/test_ast_refactor.py
@@ -1,5 +1,3 @@
-from sys import version_info
-
 import pytest
 
 import taichi as ti
@@ -515,14 +513,3 @@ def test_dict():
         foo(2)
 
 
-if version_info >= (3, 8):
-
-    @ti.test(experimental_ast_refactor=True)
-    def test_namedexpr():
-        @ti.kernel
-        def foo() -> ti.i32:
-            b = 2 + (a := 5)
-            b += a
-            return b
-
-        assert foo() == 12
